# api/openapi/swagger.yaml
basePath: /
definitions:
...
  github_com_bsonger_devflow-service_internal_platform_httpx.APIError:
    properties:
      code:
        type: string
      details:
        additionalProperties: {}
        type: object
      message:
        type: string
    type: object
...
  github_com_bsonger_devflow-service_internal_platform_httpx.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/github_com_bsonger_devflow-service_internal_platform_httpx.APIError'
    type: object
...
  github_com_bsonger_devflow-service_internal_platform_httpx.Pagination:
    properties:
      enabled:
        type: boolean
      limit:
        type: integer
      offset:
        type: integer
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
  http.CreateManifestRequestDoc:
    properties:
      application_id:
        type: string
      image_id:
        type: string
    type: object
  http.CreateReleaseRequest:
    properties:
      environment_id:
        type: string
      manifest_id:
        type: string
      type:
        type: string
    type: object
...
  http.ManifestListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/internal_manifest_transport_http.ManifestDoc'
        type: array
      pagination:
        $ref: '#/definitions/github_com_bsonger_devflow-service_internal_platform_httpx.Pagination'
    type: object
  http.ManifestResourcesResponse:
    properties:
      data:
        $ref: '#/definitions/internal_manifest_transport_http.ManifestResourcesViewDoc'
    type: object
  http.ManifestResponse:
    properties:
      data:
        $ref: '#/definitions/internal_manifest_transport_http.ManifestDoc'
    type: object
  http.ReleaseAppConfigDoc:
    properties:
      data:
        additionalProperties:
          type: string
        type: object
      files:
        items:
          $ref: '#/definitions/http.ReleaseFileDoc'
        type: array
      id:
        type: string
      mount_path:
        type: string
      name:
        type: string
      revision_id:
        type: string
      source_commit:
        type: string
      source_path:
        type: string
    type: object
  http.ReleaseDoc:
    properties:
      app_config_snapshot:
        $ref: '#/definitions/http.ReleaseAppConfigDoc'
      application_id:
        type: string
      created_at:
        type: string
      environment_id:
        type: string
      execution_intent_id:
        type: string
      external_ref:
        type: string
      id:
        type: string
      image_id:
        type: string
      manifest_id:
        type: string
      routes_snapshot:
        items:
          $ref: '#/definitions/http.ReleaseRouteDoc'
        type: array
      status:
        type: string
      steps:
        items:
          $ref: '#/definitions/http.ReleaseStepDoc'
        type: array
      type:
        type: string
      updated_at:
        type: string
    type: object
  http.ReleaseFileDoc:
    properties:
      content:
        type: string
      name:
        type: string
    type: object
  http.ReleaseListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/http.ReleaseDoc'
        type: array
      pagination:
        $ref: '#/definitions/github_com_bsonger_devflow-service_internal_platform_httpx.Pagination'
    type: object
  http.ReleaseResponse:
    properties:
      data:
        $ref: '#/definitions/http.ReleaseDoc'
    type: object
  http.ReleaseRouteDoc:
    properties:
      host:
        type: string
      id:
        type: string
      name:
        type: string
      path:
        type: string
      service_name:
        type: string
      service_port:
        type: integer
    type: object
  http.ReleaseStepDoc:
    properties:
      end_time:
        type: string
      message:
        type: string
      name:
        type: string
      progress:
        type: integer
      start_time:
        type: string
      status:
        type: string
    type: object
...
  internal_manifest_transport_http.ManifestDoc:
    properties:
      application_id:
        type: string
      artifact_digest:
        type: string
      artifact_media_type:
        type: string
      artifact_pushed_at:
        type: string
      artifact_ref:
        type: string
      artifact_repository:
        type: string
      artifact_tag:
        type: string
      created_at:
        type: string
      id:
        type: string
      image_id:
        type: string
      image_ref:
        type: string
      rendered_objects:
        items:
          $ref: '#/definitions/internal_manifest_transport_http.ManifestRenderedObjectDoc'
        type: array
      rendered_yaml:
        type: string
      services_snapshot:
        items:
          $ref: '#/definitions/internal_manifest_transport_http.ManifestServiceDoc'
        type: array
      status:
        type: string
      updated_at:
        type: string
      workload_config_snapshot:
        $ref: '#/definitions/internal_manifest_transport_http.ManifestWorkloadConfigDoc'
    type: object
  internal_manifest_transport_http.ManifestEnvVarDoc:
    properties:
      name:
        type: string
      value:
        type: string
    type: object
  internal_manifest_transport_http.ManifestGroupedResourcesDoc:
    properties:
      configmap:
        $ref: '#/definitions/internal_manifest_transport_http.ManifestRenderedResourceDoc'
      deployment:
        $ref: '#/definitions/internal_manifest_transport_http.ManifestRenderedResourceDoc'
      rollout:
        $ref: '#/definitions/internal_manifest_transport_http.ManifestRenderedResourceDoc'
      services:
        items:
          $ref: '#/definitions/internal_manifest_transport_http.ManifestRenderedResourceDoc'
        type: array
      virtualservice:
        $ref: '#/definitions/internal_manifest_transport_http.ManifestRenderedResourceDoc'
    type: object
  internal_manifest_transport_http.ManifestRenderedObjectDoc:
    properties:
      kind:
        type: string
      name:
        type: string
      namespace:
        type: string
      yaml:
        type: string
    type: object
  internal_manifest_transport_http.ManifestRenderedResourceDoc:
    properties:
      kind:
        type: string
      name:
        type: string
      namespace:
        type: string
      object:
        additionalProperties: {}
        type: object
      yaml:
        type: string
    type: object
  internal_manifest_transport_http.ManifestResourcesViewDoc:
    properties:
      application_id:
        type: string
      manifest_id:
        type: string
      rendered_objects:
        items:
          $ref: '#/definitions/internal_manifest_transport_http.ManifestRenderedResourceDoc'
        type: array
      resources:
        $ref: '#/definitions/internal_manifest_transport_http.ManifestGroupedResourcesDoc'
    type: object
  internal_manifest_transport_http.ManifestServiceDoc:
    properties:
      id:
        type: string
      name:
        type: string
      ports:
        items:
          $ref: '#/definitions/internal_manifest_transport_http.ManifestServicePortDoc'
        type: array
    type: object
  internal_manifest_transport_http.ManifestServicePortDoc:
    properties:
      name:
        type: string
      protocol:
        type: string
      service_port:
        type: integer
      target_port:
        type: integer
    type: object
  internal_manifest_transport_http.ManifestWorkloadConfigDoc:
    properties:
      env:
        items:
          $ref: '#/definitions/internal_manifest_transport_http.ManifestEnvVarDoc'
        type: array
      id:
        type: string
      name:
        type: string
      probes:
        additionalProperties: {}
        type: object
      replicas:
        type: integer
      resources:
        additionalProperties: {}
        type: object
      strategy:
        type: string
      workload_type:
        type: string
    type: object
...
info:
  contact: {}
  description: DevFlow meta-service HTTP API.
  title: DevFlow Meta Service API
  version: "1.0"
paths:
...
  /api/v1/manifests:
    get:
      parameters:
      - description: Application ID
        in: query
        name: application_id
        type: string
      - description: Image ID
        in: query
        name: image_id
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.ManifestListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_bsonger_devflow-service_internal_platform_httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_bsonger_devflow-service_internal_platform_httpx.ErrorResponse'
      summary: List manifests
      tags:
      - Manifest
    post:
      consumes:
      - application/json
      parameters:
      - description: Manifest create request
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/http.CreateManifestRequestDoc'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/http.ManifestResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_bsonger_devflow-service_internal_platform_httpx.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_bsonger_devflow-service_internal_platform_httpx.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_bsonger_devflow-service_internal_platform_httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_bsonger_devflow-service_internal_platform_httpx.ErrorResponse'
      summary: Create manifest
      tags:
      - Manifest
  /api/v1/manifests/{id}:
    delete:
      parameters:
      - description: Manifest ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_bsonger_devflow-service_internal_platform_httpx.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_bsonger_devflow-service_internal_platform_httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_bsonger_devflow-service_internal_platform_httpx.ErrorResponse'
      summary: Delete manifest
      tags:
      - Manifest
    get:
      parameters:
      - description: Manifest ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.ManifestResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_bsonger_devflow-service_internal_platform_httpx.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_bsonger_devflow-service_internal_platform_httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_bsonger_devflow-service_internal_platform_httpx.ErrorResponse'
      summary: Get manifest
      tags:
      - Manifest
  /api/v1/manifests/{id}/resources:
    get:
      parameters:
      - description: Manifest ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.ManifestResourcesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_bsonger_devflow-service_internal_platform_httpx.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_bsonger_devflow-service_internal_platform_httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_bsonger_devflow-service_internal_platform_httpx.ErrorResponse'
      summary: Get manifest frozen resources
      tags:
      - Manifest
...
  /api/v1/releases:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.ReleaseListResponse'
      summary: 获取Release列表
      tags:
      - Release
    post:
      consumes:
      - application/json
      description: 创建一个新的Release
      parameters:
      - description: Release Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/http.CreateReleaseRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/http.ReleaseResponse'
      summary: 创建Release
      tags:
      - Release
  /api/v1/releases/{id}:
    delete:
      parameters:
      - description: Release ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_bsonger_devflow-service_internal_platform_httpx.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_bsonger_devflow-service_internal_platform_httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_bsonger_devflow-service_internal_platform_httpx.ErrorResponse'
      summary: Delete release
      tags:
      - Release
    get:
      parameters:
      - description: Release ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.ReleaseResponse'
      summary: 获取Release
      tags:
      - Release
...
swagger: "2.0"